Extract idle hook, poison drips, and particle shaders/dispatch (#35, #38, #39)
H3: Add useIdleAnimation hook for the breath/sway loop every monster
model used to inline. Knight/Dragon/Werewolf/Skeleton/Goblin/Lich now
delegate the standard sin-based animation and pass monster-specific bits
(wing flap, jaw chatter, dagger wave, orb pulse, etc.) via an `extra`
callback. SlimeModel is left alone — it animates a child mesh, not the
group, so it doesn't fit the hook's contract.

M2: Add a shared <PoisonDrips> component that renders the purple
falling-circle animation used by every monster sprite. All six sprites
now pass a small `drips` config (positions, sizes, durations) instead
of inlining 8-12 lines of motion.circle each.

M3: Extract Particles.tsx and BurstParticles.tsx GLSL shaders to
src/background/shaders/{cards,burst}.ts. Pull the BurstParticles event
dispatcher (color palette, processEvent switch, lerpColor helper) out
to src/background/burstEffects.ts; the .tsx file now owns only the
buffer pool, spawn primitive, and useFrame physics. processEvent is
fully decoupled from Three.js plumbing — just takes a SpawnFn and
hero/monster world coords, so it could be unit-tested with a vi.fn()
spawn stub if desired.

No behavior changes. Bundle size unchanged (1220 vs 1221 KiB).

Closes #35, #38, #39.
