+++ src/background/BurstParticles.tsx
@@ -3,38 +3,10 @@ import { useFrame, useThree } from '@react-three/fiber';
 import * as THREE from 'three';
 import type { EffectEvent } from './useCombatEffects';
 import { portraitPositions } from '../game/portraitPositions';
+import { burstVertexShader, burstFragmentShader } from './shaders/burst';
+import { processEvent, lerpColor, type SpawnFn } from './burstEffects';
 
 const POOL_SIZE = 300;
-
-const vertexShader = /* glsl */ `
-  precision highp float;
-  attribute float aAlpha;
-  attribute float aSize;
-  attribute vec3 aColor;
-  varying float vAlpha;
-  varying vec3 vColor;
-  void main() {
-    vAlpha = aAlpha;
-    vColor = aColor;
-    vec4 mvPos = modelViewMatrix * vec4(position, 1.0);
-    gl_Position = projectionMatrix * mvPos;
-    gl_PointSize = min(aSize * (800.0 / -mvPos.z), 64.0);
-  }
-`;
-
-const fragmentShader = /* glsl */ `
-  precision highp float;
-  varying float vAlpha;
-  varying vec3 vColor;
-  void main() {
-    float dist = distance(gl_PointCoord, vec2(0.5));
-    float alpha = 1.0 - smoothstep(0.2, 0.5, dist);
-    alpha *= vAlpha;
-    if (alpha < 0.01) discard;
-    // Boost color above 1.0 for HDR bloom pickup
-    gl_FragColor = vec4(vColor * 2.5, alpha);
-  }
-`;
 
 interface ParticleData {
   active: boolean;
@@ -45,25 +17,6 @@ interface ParticleData {
   maxLife: number;
   gravity: number;
   drag: number;
-}
-
-// Color constants
-const C_GREEN = new THREE.Color('#66cc77');
-const C_WHITE = new THREE.Color('#ffffff');
-const C_RED = new THREE.Color('#ff4444');
-const C_HEAL_GREEN = new THREE.Color('#44dd44');
-const C_PURPLE = new THREE.Color('#aa44dd');
-const C_GOLD = new THREE.Color('#d4a843');
-const C_DARK_RED = new THREE.Color('#661111');
-const C_ROYAL_PURPLE = new THREE.Color('#c8a2ff');
-const C_ROYAL_GOLD = new THREE.Color('#ffd866');
-
-function lerpColor(a: THREE.Color, b: THREE.Color, t: number): THREE.Color {
-  return new THREE.Color(
-    a.r + (b.r - a.r) * t,
-    a.g + (b.g - a.g) * t,
-    a.b + (b.b - a.b) * t,
-  );
 }
 
 interface Props {
@@ -107,17 +60,17 @@ export default function BurstParticles({ effectQueue }: Props) {
     return { pool, positions, alphas, sizes, colors, geometry: geo };
   }, []);
 
-  function spawn(
-    count: number,
-    x: number, y: number, z: number,
-    velFn: (i: number) => [number, number, number],
-    color: THREE.Color,
-    size: [number, number],
-    life: [number, number],
-    gravity: number = 0,
-    drag: number = 0.98,
-    colorVariant?: THREE.Color,
-  ) {
+  const spawn: SpawnFn = (
+    count,
+    x, y, z,
+    velFn,
+    color,
+    size,
+    life,
+    gravity = 0,
+    drag = 0.98,
+    colorVariant,
+  ) => {
     for (let n = 0; n < count; n++) {
       const idx = nextIdx.current % POOL_SIZE;
       nextIdx.current++;
@@ -145,97 +98,7 @@ export default function BurstParticles({ effectQueue }: Props) {
       p.gravity = gravity;
       p.drag = drag;
     }
-  }
-
-  function processEvent(ev: EffectEvent) {
-    const isVictory = ev.label === 'victory';
-    const isDefeat = ev.label === 'defeat';
-    const isCombo = ev.label?.startsWith('Combo');
-
-    if (isVictory) {
-      // Gold explosion from center
-      spawn(100, 0, 0, 0,
-        () => [(Math.random() - 0.5) * 0.5, (Math.random() - 0.5) * 0.5, (Math.random() - 0.5) * 0.2],
-        C_GOLD, [0.15, 0.35], [1.5, 3], -0.002, 0.97, C_WHITE);
-      return;
-    }
-    if (isDefeat) {
-      // Dark red particles sinking
-      spawn(50, 0, 4, 0,
-        () => [(Math.random() - 0.5) * 0.3, -Math.random() * 0.15 - 0.05, (Math.random() - 0.5) * 0.1],
-        C_DARK_RED, [0.12, 0.25], [2, 4], 0.005, 0.99, C_RED);
-      return;
-    }
-
-    // Get portrait world positions
-    const [heroX, heroY] = screenToWorld(portraitPositions.hero.x, portraitPositions.hero.y);
-    const [monsterX, monsterY] = screenToWorld(portraitPositions.monster.x, portraitPositions.monster.y);
-
-    switch (ev.type) {
-      case 'hero-attack': {
-        const count = isCombo ? Math.min(60, 20 + ev.damage) : 30;
-        const intensity = isCombo ? 0.4 : 0.3;
-        // Burst from monster (the one being hit)
-        spawn(count, monsterX, monsterY, 0,
-          () => [(Math.random() - 0.5) * intensity, Math.random() * 0.25 + 0.1, (Math.random() - 0.5) * 0.05],
-          C_GREEN, [0.1, 0.25], [0.8, 1.8], -0.003, 0.97, C_WHITE);
-        break;
-      }
-      case 'monster-attack': {
-        // Burst from hero (the one being hit)
-        spawn(25, heroX, heroY, 0,
-          () => [(Math.random() - 0.5) * 0.2, -Math.random() * 0.2 - 0.08, (Math.random() - 0.5) * 0.05],
-          C_RED, [0.08, 0.18], [0.8, 1.5], 0.004, 0.98);
-        break;
-      }
-      case 'hero-heal': {
-        // Gentle float up from hero
-        spawn(20, heroX, heroY, 0,
-          () => [(Math.random() - 0.5) * 0.1, Math.random() * 0.1 + 0.03, (Math.random() - 0.5) * 0.03],
-          C_HEAL_GREEN, [0.08, 0.15], [1, 2], -0.001, 0.99, C_WHITE);
-        break;
-      }
-      case 'poison': {
-        // Spiral outward from monster
-        spawn(20, monsterX, monsterY, 0,
-          (i) => {
-            const angle = (i / 20) * Math.PI * 2;
-            return [Math.cos(angle) * 0.15, Math.sin(angle) * 0.15, (Math.random() - 0.5) * 0.03];
-          },
-          C_PURPLE, [0.1, 0.18], [0.8, 1.5], 0, 0.96);
-        break;
-      }
-      case 'empower': {
-        // Ring expansion from hero
-        spawn(25, heroX, heroY, 0,
-          (i) => {
-            const angle = (i / 25) * Math.PI * 2;
-            return [Math.cos(angle) * 0.18, Math.sin(angle) * 0.18, 0];
-          },
-          C_GOLD, [0.1, 0.2], [0.8, 1.5], 0, 0.97, C_WHITE);
-        break;
-      }
-      case 'face-card': {
-        // Royal Awakening — tier-scaled particle burst from hero
-        const isAwakens = ev.label?.includes('Awakens');
-        const isRises = ev.label?.includes('Rises');
-        const count = isAwakens ? 40 : isRises ? 25 : 15;
-        const speed = isAwakens ? 0.25 : isRises ? 0.18 : 0.1;
-        const baseColor = isAwakens ? C_ROYAL_GOLD : C_ROYAL_PURPLE;
-        const variantColor = isAwakens ? C_WHITE : C_ROYAL_GOLD;
-        const sizeRange: [number, number] = isAwakens ? [0.12, 0.3] : isRises ? [0.1, 0.22] : [0.06, 0.14];
-        const lifeRange: [number, number] = isAwakens ? [1.2, 2.5] : isRises ? [0.8, 1.8] : [0.6, 1.2];
-        spawn(count, heroX, heroY, 0,
-          (n) => {
-            const angle = (n / count) * Math.PI * 2 + Math.random() * 0.3;
-            const r = speed * (0.7 + Math.random() * 0.6);
-            return [Math.cos(angle) * r, Math.sin(angle) * r + 0.05, (Math.random() - 0.5) * 0.05];
-          },
-          baseColor, sizeRange, lifeRange, -0.001, 0.97, variantColor);
-        break;
-      }
-    }
-  }
+  };
 
   // Delayed events waiting for attack wind-up to finish
   const delayedEvents = useRef<{ ev: EffectEvent; fireAt: number }[]>([]);
@@ -252,6 +115,13 @@ export default function BurstParticles({ effectQueue }: Props) {
     // Process queued events — delay attack particles until wind-up ends
     const queue = effectQueue.current;
     if (!queue) return;
+
+    const ctx = {
+      spawn,
+      heroPos: screenToWorld(portraitPositions.hero.x, portraitPositions.hero.y),
+      monsterPos: screenToWorld(portraitPositions.monster.x, portraitPositions.monster.y),
+    };
+
     while (queue.length > 0) {
       const ev = queue.shift()!;
       const isVictory = ev.label === 'victory';
@@ -261,7 +131,7 @@ export default function BurstParticles({ effectQueue }: Props) {
       } else if (!isVictory && !isDefeat && ev.type === 'monster-attack') {
         delayedEvents.current.push({ ev, fireAt: now + MONSTER_ANTICIPATE });
       } else {
-        processEvent(ev);
+        processEvent(ev, ctx);
       }
     }
 
@@ -269,7 +139,7 @@ export default function BurstParticles({ effectQueue }: Props) {
     const pending = delayedEvents.current;
     for (let i = pending.length - 1; i >= 0; i--) {
       if (now >= pending[i].fireAt) {
-        processEvent(pending[i].ev);
+        processEvent(pending[i].ev, ctx);
         pending.splice(i, 1);
       }
     }
@@ -313,8 +183,8 @@ export default function BurstParticles({ effectQueue }: Props) {
   return (
     <points ref={meshRef} geometry={geometry}>
       <shaderMaterial
-        vertexShader={vertexShader}
-        fragmentShader={fragmentShader}
+        vertexShader={burstVertexShader}
+        fragmentShader={burstFragmentShader}
         uniforms={{}}
         transparent
         depthWrite={false}
