+++ src/combat/sprites/SkeletonSprite.tsx
@@ -1,4 +1,4 @@
-import { motion } from 'framer-motion';
+import PoisonDrips from './PoisonDrips';
 
 interface Props {
   poisoned: boolean;
@@ -46,16 +46,13 @@ export default function SkeletonSprite({ poisoned }: Props) {
       <line x1="32" y1="48" x2="24" y2="60" stroke="#d0c8b8" strokeWidth="2" strokeLinecap="round" />
       <line x1="32" y1="48" x2="40" y2="60" stroke="#d0c8b8" strokeWidth="2" strokeLinecap="round" />
 
-      {poisoned && (
-        <>
-          <motion.circle cx="26" cy="50" r="2" fill="#8a44bb"
-            animate={{ cy: [50, 58], opacity: [0.8, 0] }}
-            transition={{ duration: 1.2, repeat: Infinity }} />
-          <motion.circle cx="38" cy="48" r="1.5" fill="#aa55dd"
-            animate={{ cy: [48, 56], opacity: [0.7, 0] }}
-            transition={{ duration: 1, repeat: Infinity, delay: 0.4 }} />
-        </>
-      )}
+      <PoisonDrips
+        poisoned={poisoned}
+        drips={[
+          { cx: 26, cy: 50, r: 2, duration: 1.2 },
+          { cx: 38, cy: 48, r: 1.5, duration: 1, opacity: 0.7 },
+        ]}
+      />
     </svg>
   );
 }
