+++ src/combat/sprites/SlimeSprite.tsx
@@ -1,4 +1,4 @@
-import { motion } from 'framer-motion';
+import PoisonDrips from './PoisonDrips';
 
 interface Props {
   poisoned: boolean;
@@ -27,16 +27,13 @@ export default function SlimeSprite({ poisoned }: Props) {
       <ellipse cx="14" cy="54" rx="3" ry="2" fill="#44cc44" opacity="0.5" />
       <ellipse cx="50" cy="55" rx="2.5" ry="1.5" fill="#44cc44" opacity="0.4" />
 
-      {poisoned && (
-        <>
-          <motion.circle cx="24" cy="50" r="2" fill="#8a44bb"
-            animate={{ cy: [50, 58], opacity: [0.8, 0] }}
-            transition={{ duration: 1.2, repeat: Infinity }} />
-          <motion.circle cx="40" cy="48" r="1.5" fill="#aa55dd"
-            animate={{ cy: [48, 56], opacity: [0.7, 0] }}
-            transition={{ duration: 1, repeat: Infinity, delay: 0.4 }} />
-        </>
-      )}
+      <PoisonDrips
+        poisoned={poisoned}
+        drips={[
+          { cx: 24, cy: 50, r: 2, duration: 1.2 },
+          { cx: 40, cy: 48, r: 1.5, duration: 1, opacity: 0.7 },
+        ]}
+      />
     </svg>
   );
 }
